requirements.txt
langchain
langchain_groq
langchain-community
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain-huggingface